+++ R/g_forest.R
@@ -53,7 +53,6 @@
 #' @examples
 #' library(dplyr)
 #' library(forcats)
-#' library(nestcolor)
 #'
 #' adrs <- tern_ex_adrs
 #' n_records <- 20