+++ R/incidence_rate.R
@@ -19,8 +19,9 @@
 #' @param n_events (`string`)\cr name of integer variable indicating whether an event has been observed (1) or not (0).
 #' @param id_var (`string`)\cr name of variable used as patient identifier if `"n_unique"` is included in `.stats`.
 #'   Defaults to `"USUBJID"`.
-#' @param .stats (`character`)\cr statistics to select for the table. Run `get_stats("estimate_incidence_rate")`
-#'   to see available statistics for this function.
+#' @param .stats (`character`)\cr statistics to select for the table.
+#'
+#'   Options are: ``r shQuote(get_stats("estimate_incidence_rate"))``
 #' @param summarize (`flag`)\cr whether the function should act as an analyze function (`summarize = FALSE`), or a
 #'   summarize function (`summarize = TRUE`). Defaults to `FALSE`.
 #' @param label_fmt (`string`)\cr how labels should be formatted after a row split occurs if `summarize = TRUE`. The