+++ R/summarize_change.R
@@ -12,8 +12,9 @@
 #' or the change from baseline values (post-baseline) are then summarized.
 #'
 #' @inheritParams argument_convention
-#' @param .stats (`character`)\cr statistics to select for the table. Run `get_stats("analyze_vars_numeric)`
-#'   to see available statistics for this function.
+#' @param .stats (`character`)\cr statistics to select for the table.
+#'
+#'   Options are: ``r shQuote(get_stats("analyze_vars_numeric"))``
 #'
 #' @name summarize_change
 #' @order 1
@@ -28,11 +29,12 @@ NULL
 #'   an error will be thrown.
 #'
 #' @keywords internal
-s_change_from_baseline <- function(df,
-                                   .var,
-                                   variables,
-                                   na.rm = TRUE, # nolint
-                                   ...) {
+s_change_from_baseline <- function(df, ...) {
+  # s_summary should get na.rm
+  args_list <- list(...)
+  .var <- args_list[[".var"]]
+  variables <- args_list[["variables"]]
+
   checkmate::assert_numeric(df[[variables$value]])
   checkmate::assert_numeric(df[[.var]])
   checkmate::assert_logical(df[[variables$baseline_flag]])
@@ -47,7 +49,7 @@ s_change_from_baseline <- function(df,
   if (is.logical(combined) && identical(length(combined), 0L)) {
     combined <- numeric(0)
   }
-  s_summary(combined, na.rm = na.rm, ...)
+  s_summary(combined, ...)
 }
 
 #' @describeIn summarize_change Formatted analysis function which is used as `afun` in `summarize_change()`.
@@ -56,25 +58,56 @@ s_change_from_baseline <- function(df,
 #' * `a_change_from_baseline()` returns the corresponding list with formatted [rtables::CellValue()].
 #'
 #' @keywords internal
-a_change_from_baseline <- make_afun(
-  s_change_from_baseline,
-  .formats = c(
-    n = "xx",
-    mean_sd = "xx.xx (xx.xx)",
-    mean_se = "xx.xx (xx.xx)",
-    median = "xx.xx",
-    range = "xx.xx - xx.xx",
-    mean_ci = "(xx.xx, xx.xx)",
-    median_ci = "(xx.xx, xx.xx)",
-    mean_pval = "xx.xx"
-  ),
-  .labels = c(
-    mean_sd = "Mean (SD)",
-    mean_se = "Mean (SE)",
-    median = "Median",
-    range = "Min - Max"
+a_change_from_baseline <- function(df,
+                                   ...,
+                                   .stats = NULL,
+                                   .formats = NULL,
+                                   .labels = NULL,
+                                   .indent_mods = NULL) {
+  # Check if there are user-defined functions
+  default_and_custom_stats_list <- .split_std_from_custom_stats(.stats)
+  .stats <- default_and_custom_stats_list$default_stats
+  custom_stat_functions <- default_and_custom_stats_list$custom_stats
+
+  # Adding automatically extra parameters to the statistic function (see ?rtables::additional_fun_params)
+  extra_afun_params <- retrieve_extra_afun_params(
+    names(list(...)$.additional_fun_parameters)
   )
-)
+  x_stats <- .apply_stat_functions(
+    default_stat_fnc = s_change_from_baseline,
+    custom_stat_fnc_list = custom_stat_functions,
+    args_list = c(
+      df = list(df),
+      extra_afun_params,
+      list(...)
+    )
+  )
+
+  # Fill in with formatting defaults if needed
+  .stats <- c(
+    get_stats("analyze_vars_numeric", stats_in = .stats),
+    names(custom_stat_functions) # Additional stats from custom functions
+  )
+  .formats <- get_formats_from_stats(.stats, .formats)
+  .labels <- get_labels_from_stats(.stats, .labels)
+  .indent_mods <- get_indents_from_stats(.stats, .indent_mods)
+
+  # Auto format handling
+  .formats <- apply_auto_formatting(
+    .formats,
+    x_stats,
+    extra_afun_params$.df_row,
+    extra_afun_params$.var
+  )
+
+  in_rows(
+    .list = x_stats[.stats],
+    .formats = .formats,
+    .names = names(.labels),
+    .labels = .labels,
+    .indent_mods = .indent_mods
+  )
+}
 
 #' @describeIn summarize_change Layout-creating function which can take statistics function arguments
 #'   and additional format arguments. This function is a wrapper for [rtables::analyze()].
@@ -90,7 +123,7 @@ a_change_from_baseline <- make_afun(
 #' @examples
 #' library(dplyr)
 #'
-#' ## Fabricate dataset
+#' # Fabricate dataset
 #' dta_test <- data.frame(
 #'   USUBJID = rep(1:6, each = 3),
 #'   AVISIT = rep(paste0("V", 1:3), 6),
@@ -118,31 +151,60 @@ a_change_from_baseline <- make_afun(
 summarize_change <- function(lyt,
                              vars,
                              variables,
+                             var_labels = vars,
                              na_str = default_na_str(),
+                             na_rm = TRUE,
                              nested = TRUE,
-                             ...,
+                             show_labels = "default",
                              table_names = vars,
+                             section_div = NA_character_,
+                             ...,
                              .stats = c("n", "mean_sd", "median", "range"),
-                             .formats = NULL,
-                             .labels = NULL,
+                             .formats = c(
+                               n = "xx",
+                               mean_sd = "xx.xx (xx.xx)",
+                               mean_se = "xx.xx (xx.xx)",
+                               median = "xx.xx",
+                               range = "xx.xx - xx.xx",
+                               mean_ci = "(xx.xx, xx.xx)",
+                               median_ci = "(xx.xx, xx.xx)",
+                               mean_pval = "xx.xx"
+                             ),
+                             .labels = c(
+                               mean_sd = "Mean (SD)",
+                               mean_se = "Mean (SE)",
+                               median = "Median",
+                               range = "Min - Max"
+                             ),
                              .indent_mods = NULL) {
-  extra_args <- list(variables = variables, ...)
+  # Extra args must contain .stats, .formats, .labels, .indent_mods - sent to the analysis level
+  extra_args <- list(".stats" = .stats)
+  if (!is.null(.formats)) extra_args[[".formats"]] <- .formats
+  if (!is.null(.labels)) extra_args[[".labels"]] <- .labels
+  if (!is.null(.indent_mods)) extra_args[[".indent_mods"]] <- .indent_mods
 
-  afun <- make_afun(
-    a_change_from_baseline,
-    .stats = .stats,
-    .formats = .formats,
-    .labels = .labels,
-    .indent_mods = .indent_mods
+  # Adding additional arguments to the analysis function (depends on the specific call)
+  extra_args <- c(extra_args, "variables" = list(variables), ...)
+
+  # Adding all additional information from layout to analysis functions (see ?rtables::additional_fun_params)
+  extra_args[[".additional_fun_parameters"]] <- get_additional_afun_params(add_alt_df = FALSE)
+  formals(a_change_from_baseline) <- c(
+    formals(a_change_from_baseline),
+    extra_args[[".additional_fun_parameters"]]
   )
 
+  # Main analysis call - Nothing with .* -> these should be dedicated to the analysis function
   analyze(
-    lyt,
-    vars,
-    afun = afun,
+    lyt = lyt,
+    vars = vars,
+    var_labels = var_labels,
+    afun = a_change_from_baseline,
     na_str = na_str,
     nested = nested,
     extra_args = extra_args,
-    table_names = table_names
+    inclNAs = na_rm, # adds na.rm = TRUE to the analysis function
+    show_labels = show_labels,
+    table_names = table_names,
+    section_div = section_div
   )
 }