+++ R/summarize_glm_count.R
@@ -14,8 +14,9 @@
 #' @param rate_mean_method (`character(1)`)\cr method used to estimate the mean odds ratio. Defaults to `emmeans`.
 #'   see details for more information.
 #' @param scale (`numeric(1)`)\cr linear scaling factor for rate and confidence intervals. Defaults to `1`.
-#' @param .stats (`character`)\cr statistics to select for the table. Run `get_stats("summarize_glm_count")`
-#'   to see available statistics for this function.
+#' @param .stats (`character`)\cr statistics to select for the table.
+#'
+#'   Options are: ``r shQuote(get_stats("summarize_glm_count"))``
 #'
 #' @details
 #' `summarize_glm_count()` uses `s_glm_count()` to calculate the statistics for the table. This